backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
torch==2.4.1
torchvision==0.19.1
scikit-learn==1.5.2
joblib==1.5.3
pillow==12.1.1
pandas==3.0.1
numpy==1.26.4
tensorflow-cpu
python-dotenv
requests
gunicorn
